group {rex}	R Documentation

Create a grouped expression

Description

This is similar to capture except that it does not store the value of the group. Best used when you want to combine several parts together and do not reference or extract the grouped value later.

Usage

group(...)

Arguments

...
shortcuts, R variables, text, or other rex functions.

See Also

capture for grouping with capturing. Perl 5 Extended Patterns http://perldoc.perl.org/perlre.html#Extended-Patterns

Other rex: %or%(), capture(), character_class(), counts, lookarounds, not(), rex(), shortcuts, wildcards

[Package rex version 1.2.0 Index]