register_shortcuts {rex}	R Documentation

Register the Rex shortcuts

Description

If you are using rex in another package you need to call this function to register all of the rex shortcuts so that spurious NOTEs about global variables being generated during R CMD check.

Usage

register_shortcuts(pkg_name)

Arguments

pkg_name
the package to register the shortcuts in

[Package rex version 1.2.1 Index]