uptakeGuess

Function to help find good initial parameters by guessing the initial uptake using a generalised inverse appraoach

uptakeGuess(res, numRep, numPeptides, numtimepoints, R, C, phi = 1)

Arguments

res

An object of class DataFrame containing the results of the hdx-ms experiment

numRep

The number of replicates

numPeptides

The number of peptides

numtimepoints

The number of timepoints

R

The number of residues

C

A matrix indicating the position of the residues in the peptides.

phi

The maximum uptake possible. Default is 1.

Value

A matrix of intial uptake guesses

Examples

require(ReX)
require(dplyr)
data(BRD4_apo)

BRD4_apo <- BRD4_apo %>% filter(End < 100)
numTimepoints <- length(unique(BRD4_apo$Exposure))
Timepoints <- unique(BRD4_apo$Exposure)
numPeptides <- length(unique(BRD4_apo$Sequence))
BRD4_apo <- DataFrame(BRD4_apo)
C <- coverageHeatmap(res = BRD4_apo, plot = FALSE)

uptakeGuess(BRD4_apo,
            numRep = 3,
            numPeptides =  numPeptides,
            numtimepoints =  numTimepoints,
            R =  99,
            C = C,
            phi = 0.92)
#> [[1]]
#>        [,1]         [,2]         [,3]         [,4]         [,5]         [,6]
#>   [1,]    0 9.925906e-17 2.686649e-16 1.862456e-16 3.598183e-16 5.500804e-16
#>   [2,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>   [3,]    0 8.892876e-01 9.200000e-01 9.200000e-01 8.898015e-01 9.200000e-01
#>   [4,]    0 8.892876e-01 9.200000e-01 9.200000e-01 8.898015e-01 9.200000e-01
#>   [5,]    0 8.892876e-01 9.200000e-01 9.200000e-01 8.898015e-01 9.200000e-01
#>   [6,]    0 8.892876e-01 9.200000e-01 9.200000e-01 8.898015e-01 9.200000e-01
#>   [7,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>   [8,]    0 8.633934e-01 7.769168e-01 6.898504e-01 6.817070e-01 7.055939e-01
#>   [9,]    0 8.633934e-01 7.769168e-01 6.898504e-01 6.817070e-01 7.055939e-01
#>  [10,]    0 8.633934e-01 7.769168e-01 6.898504e-01 6.817070e-01 7.055939e-01
#>  [11,]    0 8.633934e-01 7.769168e-01 6.898504e-01 6.817070e-01 7.055939e-01
#>  [12,]    0 8.633934e-01 7.769168e-01 6.898504e-01 6.817070e-01 7.055939e-01
#>  [13,]    0 8.633934e-01 7.769168e-01 6.898504e-01 6.817070e-01 7.055939e-01
#>  [14,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [15,]    0 1.554270e-17 2.085759e-17 3.002036e-17 2.984396e-17 3.899956e-17
#>  [16,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [17,]    0 7.798889e-01 8.151160e-01 7.712452e-01 7.961978e-01 8.193563e-01
#>  [18,]    0 7.798889e-01 8.151160e-01 7.712452e-01 7.961978e-01 8.193563e-01
#>  [19,]    0 7.798889e-01 8.151160e-01 7.712452e-01 7.961978e-01 8.193563e-01
#>  [20,]    0 7.798889e-01 8.151160e-01 7.712452e-01 7.961978e-01 8.193563e-01
#>  [21,]    0 7.798889e-01 8.151160e-01 7.712452e-01 7.961978e-01 8.193563e-01
#>  [22,]    0 7.798889e-01 8.151160e-01 7.712452e-01 7.961978e-01 8.193563e-01
#>  [23,]    0 7.798889e-01 8.151160e-01 7.712452e-01 7.961978e-01 8.193563e-01
#>  [24,]    0 7.798889e-01 8.151160e-01 7.712452e-01 7.961978e-01 8.193563e-01
#>  [25,]    0 7.798889e-01 8.151160e-01 7.712452e-01 7.961978e-01 8.193563e-01
#>  [26,]    0 7.798889e-01 8.151160e-01 7.712452e-01 7.961978e-01 8.193563e-01
#>  [27,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [28,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [29,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [30,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [31,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [32,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [33,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [34,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [35,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [36,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [37,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [38,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [39,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [40,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [41,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [42,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [43,]    0 8.524781e-01 9.194326e-01 8.591736e-01 8.356827e-01 8.685926e-01
#>  [44,]    0 8.524781e-01 9.194326e-01 8.591736e-01 8.356827e-01 8.685926e-01
#>  [45,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [46,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [47,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [48,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [49,]    0 8.524781e-01 9.194326e-01 8.591736e-01 8.356827e-01 8.685926e-01
#>  [50,]    0 8.524781e-01 9.194326e-01 8.591736e-01 8.356827e-01 8.685926e-01
#>  [51,]    0 8.524781e-01 9.194326e-01 8.591736e-01 8.356827e-01 8.685926e-01
#>  [52,]    0 8.524781e-01 9.194326e-01 8.591736e-01 8.356827e-01 8.685926e-01
#>  [53,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [54,]    0 8.524781e-01 9.194326e-01 8.591736e-01 8.356827e-01 8.685926e-01
#>  [55,]    0 8.524781e-01 9.194326e-01 8.591736e-01 8.356827e-01 8.685926e-01
#>  [56,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [57,]    0 8.524781e-01 9.194326e-01 8.591736e-01 8.356827e-01 8.685926e-01
#>  [58,]    0 8.524781e-01 9.194326e-01 8.591736e-01 8.356827e-01 8.685926e-01
#>  [59,]    0 8.524781e-01 9.194326e-01 8.591736e-01 8.356827e-01 8.685926e-01
#>  [60,]    0 8.524781e-01 9.194326e-01 8.591736e-01 8.356827e-01 8.685926e-01
#>  [61,]    0 8.524781e-01 9.194326e-01 8.591736e-01 8.356827e-01 8.685926e-01
#>  [62,]    0 8.524781e-01 9.194326e-01 8.591736e-01 8.356827e-01 8.685926e-01
#>  [63,]    0 8.524781e-01 9.194326e-01 8.591736e-01 8.356827e-01 8.685926e-01
#>  [64,]    0 0.000000e+00 1.686060e-01 1.316468e-01 5.826832e-01 8.425987e-01
#>  [65,]    0 0.000000e+00 1.686060e-01 1.316468e-01 5.826832e-01 8.425987e-01
#>  [66,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [67,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [68,]    0 3.258963e-04 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [69,]    0 3.258963e-04 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [70,]    0 0.000000e+00 5.530497e-02 4.031163e-02 0.000000e+00 9.872417e-02
#>  [71,]    0 5.561749e-02 9.082274e-02 2.502774e-01 3.780328e-01 5.185659e-01
#>  [72,]    0 5.561749e-02 9.082274e-02 2.502774e-01 3.780328e-01 5.185659e-01
#>  [73,]    0 5.529160e-02 1.183914e-01 3.217333e-01 4.301564e-01 5.836470e-01
#>  [74,]    0 5.529160e-02 1.183914e-01 3.217333e-01 4.301564e-01 5.836470e-01
#>  [75,]    0 6.324777e-02 3.551777e-02 2.099658e-01 4.119406e-01 4.198417e-01
#>  [76,]    0 6.324777e-02 3.551777e-02 2.099658e-01 4.119406e-01 4.198417e-01
#>  [77,]    0 6.083872e-01 7.865208e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [78,]    0 6.083872e-01 7.865208e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [79,]    0 1.915543e-01 1.451382e-01 8.172445e-01 7.032213e-01 5.653994e-01
#>  [80,]    0 1.011330e-01 1.472059e-01 3.554958e-01 5.090334e-01 7.442837e-01
#>  [81,]    0 4.111815e-01 5.995924e-01 8.322129e-01 9.084720e-01 9.200000e-01
#>  [82,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [83,]    0 4.111815e-01 5.995924e-01 8.322129e-01 9.084720e-01 9.200000e-01
#>  [84,]    0 1.240986e-01 2.004669e-01 3.364981e-01 4.468742e-01 5.459144e-01
#>  [85,]    0 1.240986e-01 2.004669e-01 3.364981e-01 4.468742e-01 5.459144e-01
#>  [86,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [87,]    0 1.240986e-01 2.004669e-01 3.364981e-01 4.468742e-01 5.459144e-01
#>  [88,]    0 1.240986e-01 2.004669e-01 3.364981e-01 4.468742e-01 5.459144e-01
#>  [89,]    0 1.939424e-01 4.386837e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [90,]    0 1.939424e-01 4.386837e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [91,]    0 1.939424e-01 4.386837e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [92,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [93,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [94,]    0 2.796829e-01 5.079509e-01 8.911930e-01 9.014031e-01 9.123424e-01
#>  [95,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [96,]    0 2.796829e-01 5.079509e-01 8.911930e-01 9.014031e-01 9.123424e-01
#>  [97,]    0 2.796829e-01 5.079509e-01 8.911930e-01 9.014031e-01 9.123424e-01
#>  [98,]    0 2.796829e-01 5.079509e-01 8.911930e-01 9.014031e-01 9.123424e-01
#>  [99,]    0 2.796829e-01 5.079509e-01 8.911930e-01 9.014031e-01 9.123424e-01
#> 
#> [[2]]
#>        [,1]         [,2]         [,3]         [,4]         [,5]         [,6]
#>   [1,]    0 1.394179e-16 2.486679e-16 1.775982e-16 3.546193e-16 3.920146e-16
#>   [2,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>   [3,]    0 9.200000e-01 9.200000e-01 8.252618e-01 8.946036e-01 8.941222e-01
#>   [4,]    0 9.200000e-01 9.200000e-01 8.252618e-01 8.946036e-01 8.941222e-01
#>   [5,]    0 9.200000e-01 9.200000e-01 8.252618e-01 8.946036e-01 8.941222e-01
#>   [6,]    0 9.200000e-01 9.200000e-01 8.252618e-01 8.946036e-01 8.941222e-01
#>   [7,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>   [8,]    0 8.114006e-01 8.254846e-01 6.893098e-01 6.845798e-01 7.047389e-01
#>   [9,]    0 8.114006e-01 8.254846e-01 6.893098e-01 6.845798e-01 7.047389e-01
#>  [10,]    0 8.114006e-01 8.254846e-01 6.893098e-01 6.845798e-01 7.047389e-01
#>  [11,]    0 8.114006e-01 8.254846e-01 6.893098e-01 6.845798e-01 7.047389e-01
#>  [12,]    0 8.114006e-01 8.254846e-01 6.893098e-01 6.845798e-01 7.047389e-01
#>  [13,]    0 8.114006e-01 8.254846e-01 6.893098e-01 6.845798e-01 7.047389e-01
#>  [14,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [15,]    0 1.441179e-17 2.202620e-17 2.863447e-17 3.021718e-17 4.223080e-17
#>  [16,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [17,]    0 7.687989e-01 8.164207e-01 7.828245e-01 8.095710e-01 8.273476e-01
#>  [18,]    0 7.687989e-01 8.164207e-01 7.828245e-01 8.095710e-01 8.273476e-01
#>  [19,]    0 7.687989e-01 8.164207e-01 7.828245e-01 8.095710e-01 8.273476e-01
#>  [20,]    0 7.687989e-01 8.164207e-01 7.828245e-01 8.095710e-01 8.273476e-01
#>  [21,]    0 7.687989e-01 8.164207e-01 7.828245e-01 8.095710e-01 8.273476e-01
#>  [22,]    0 7.687989e-01 8.164207e-01 7.828245e-01 8.095710e-01 8.273476e-01
#>  [23,]    0 7.687989e-01 8.164207e-01 7.828245e-01 8.095710e-01 8.273476e-01
#>  [24,]    0 7.687989e-01 8.164207e-01 7.828245e-01 8.095710e-01 8.273476e-01
#>  [25,]    0 7.687989e-01 8.164207e-01 7.828245e-01 8.095710e-01 8.273476e-01
#>  [26,]    0 7.687989e-01 8.164207e-01 7.828245e-01 8.095710e-01 8.273476e-01
#>  [27,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [28,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [29,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [30,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [31,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [32,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [33,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [34,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [35,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [36,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [37,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [38,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [39,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [40,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [41,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [42,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [43,]    0 8.207031e-01 9.180708e-01 8.506624e-01 8.292143e-01 9.045664e-01
#>  [44,]    0 8.207031e-01 9.180708e-01 8.506624e-01 8.292143e-01 9.045664e-01
#>  [45,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [46,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [47,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [48,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [49,]    0 8.207031e-01 9.180708e-01 8.506624e-01 8.292143e-01 9.045664e-01
#>  [50,]    0 8.207031e-01 9.180708e-01 8.506624e-01 8.292143e-01 9.045664e-01
#>  [51,]    0 8.207031e-01 9.180708e-01 8.506624e-01 8.292143e-01 9.045664e-01
#>  [52,]    0 8.207031e-01 9.180708e-01 8.506624e-01 8.292143e-01 9.045664e-01
#>  [53,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [54,]    0 8.207031e-01 9.180708e-01 8.506624e-01 8.292143e-01 9.045664e-01
#>  [55,]    0 8.207031e-01 9.180708e-01 8.506624e-01 8.292143e-01 9.045664e-01
#>  [56,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [57,]    0 8.207031e-01 9.180708e-01 8.506624e-01 8.292143e-01 9.045664e-01
#>  [58,]    0 8.207031e-01 9.180708e-01 8.506624e-01 8.292143e-01 9.045664e-01
#>  [59,]    0 8.207031e-01 9.180708e-01 8.506624e-01 8.292143e-01 9.045664e-01
#>  [60,]    0 8.207031e-01 9.180708e-01 8.506624e-01 8.292143e-01 9.045664e-01
#>  [61,]    0 8.207031e-01 9.180708e-01 8.506624e-01 8.292143e-01 9.045664e-01
#>  [62,]    0 8.207031e-01 9.180708e-01 8.506624e-01 8.292143e-01 9.045664e-01
#>  [63,]    0 8.207031e-01 9.180708e-01 8.506624e-01 8.292143e-01 9.045664e-01
#>  [64,]    0 0.000000e+00 1.123352e-01 2.684509e-01 6.368723e-01 8.395424e-01
#>  [65,]    0 0.000000e+00 1.123352e-01 2.684509e-01 6.368723e-01 8.395424e-01
#>  [66,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [67,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [68,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [69,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [70,]    0 0.000000e+00 8.255280e-02 1.259698e-02 0.000000e+00 0.000000e+00
#>  [71,]    0 4.738926e-02 8.208924e-02 2.815931e-01 3.888801e-01 6.055377e-01
#>  [72,]    0 4.738926e-02 8.208924e-02 2.815931e-01 3.888801e-01 6.055377e-01
#>  [73,]    0 7.336701e-02 1.380696e-01 3.609717e-01 4.289109e-01 6.770972e-01
#>  [74,]    0 7.336701e-02 1.380696e-01 3.609717e-01 4.289109e-01 6.770972e-01
#>  [75,]    0 4.825538e-02 0.000000e+00 2.689961e-01 4.445113e-01 6.156718e-01
#>  [76,]    0 4.825538e-02 0.000000e+00 2.689961e-01 4.445113e-01 6.156718e-01
#>  [77,]    0 4.705249e-01 8.115575e-01 9.084535e-01 9.200000e-01 9.200000e-01
#>  [78,]    0 4.705249e-01 8.115575e-01 9.084535e-01 9.200000e-01 9.200000e-01
#>  [79,]    0 3.146768e-01 2.340969e-01 7.164732e-01 7.330415e-01 8.945444e-01
#>  [80,]    0 7.707927e-02 2.635607e-01 4.483945e-01 5.230362e-01 6.832286e-01
#>  [81,]    0 4.005646e-01 5.243330e-01 7.787050e-01 8.955439e-01 9.200000e-01
#>  [82,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [83,]    0 4.005646e-01 5.243330e-01 7.787050e-01 8.955439e-01 9.200000e-01
#>  [84,]    0 1.199222e-01 1.950973e-01 3.588716e-01 4.423995e-01 5.387549e-01
#>  [85,]    0 1.199222e-01 1.950973e-01 3.588716e-01 4.423995e-01 5.387549e-01
#>  [86,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [87,]    0 1.199222e-01 1.950973e-01 3.588716e-01 4.423995e-01 5.387549e-01
#>  [88,]    0 1.199222e-01 1.950973e-01 3.588716e-01 4.423995e-01 5.387549e-01
#>  [89,]    0 2.223014e-01 4.824473e-01 9.146112e-01 9.200000e-01 9.200000e-01
#>  [90,]    0 2.223014e-01 4.824473e-01 9.146112e-01 9.200000e-01 9.200000e-01
#>  [91,]    0 2.223014e-01 4.824473e-01 9.146112e-01 9.200000e-01 9.200000e-01
#>  [92,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [93,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [94,]    0 3.227111e-01 5.064923e-01 8.769719e-01 9.035910e-01 9.200000e-01
#>  [95,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [96,]    0 3.227111e-01 5.064923e-01 8.769719e-01 9.035910e-01 9.200000e-01
#>  [97,]    0 3.227111e-01 5.064923e-01 8.769719e-01 9.035910e-01 9.200000e-01
#>  [98,]    0 3.227111e-01 5.064923e-01 8.769719e-01 9.035910e-01 9.200000e-01
#>  [99,]    0 3.227111e-01 5.064923e-01 8.769719e-01 9.035910e-01 9.200000e-01
#> 
#> [[3]]
#>        [,1]         [,2]         [,3]         [,4]         [,5]         [,6]
#>   [1,]    0 1.874745e-16 2.448976e-16 1.635107e-16 2.412128e-16 4.503866e-16
#>   [2,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>   [3,]    0 9.200000e-01 9.200000e-01 9.200000e-01 8.816935e-01 9.200000e-01
#>   [4,]    0 9.200000e-01 9.200000e-01 9.200000e-01 8.816935e-01 9.200000e-01
#>   [5,]    0 9.200000e-01 9.200000e-01 9.200000e-01 8.816935e-01 9.200000e-01
#>   [6,]    0 9.200000e-01 9.200000e-01 9.200000e-01 8.816935e-01 9.200000e-01
#>   [7,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>   [8,]    0 8.090815e-01 7.398375e-01 7.166356e-01 6.480050e-01 7.372438e-01
#>   [9,]    0 8.090815e-01 7.398375e-01 7.166356e-01 6.480050e-01 7.372438e-01
#>  [10,]    0 8.090815e-01 7.398375e-01 7.166356e-01 6.480050e-01 7.372438e-01
#>  [11,]    0 8.090815e-01 7.398375e-01 7.166356e-01 6.480050e-01 7.372438e-01
#>  [12,]    0 8.090815e-01 7.398375e-01 7.166356e-01 6.480050e-01 7.372438e-01
#>  [13,]    0 8.090815e-01 7.398375e-01 7.166356e-01 6.480050e-01 7.372438e-01
#>  [14,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.126669e-01 9.200000e-01
#>  [15,]    0 1.645444e-17 1.889732e-17 2.835663e-17 2.829513e-17 3.454346e-17
#>  [16,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.126669e-01 9.200000e-01
#>  [17,]    0 7.670049e-01 7.953823e-01 7.895111e-01 8.456136e-01 8.159315e-01
#>  [18,]    0 7.670049e-01 7.953823e-01 7.895111e-01 8.456136e-01 8.159315e-01
#>  [19,]    0 7.670049e-01 7.953823e-01 7.895111e-01 8.456136e-01 8.159315e-01
#>  [20,]    0 7.670049e-01 7.953823e-01 7.895111e-01 8.456136e-01 8.159315e-01
#>  [21,]    0 7.670049e-01 7.953823e-01 7.895111e-01 8.456136e-01 8.159315e-01
#>  [22,]    0 7.670049e-01 7.953823e-01 7.895111e-01 8.456136e-01 8.159315e-01
#>  [23,]    0 7.670049e-01 7.953823e-01 7.895111e-01 8.456136e-01 8.159315e-01
#>  [24,]    0 7.670049e-01 7.953823e-01 7.895111e-01 8.456136e-01 8.159315e-01
#>  [25,]    0 7.670049e-01 7.953823e-01 7.895111e-01 8.456136e-01 8.159315e-01
#>  [26,]    0 7.670049e-01 7.953823e-01 7.895111e-01 8.456136e-01 8.159315e-01
#>  [27,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [28,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [29,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [30,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [31,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [32,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [33,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [34,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [35,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [36,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [37,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [38,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [39,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [40,]    0 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01 9.200000e-01
#>  [41,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [42,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [43,]    0 8.506624e-01 8.806217e-01 8.410164e-01 8.421512e-01 8.631454e-01
#>  [44,]    0 8.506624e-01 8.806217e-01 8.410164e-01 8.421512e-01 8.631454e-01
#>  [45,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [46,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [47,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [48,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [49,]    0 8.506624e-01 8.806217e-01 8.410164e-01 8.421512e-01 8.631454e-01
#>  [50,]    0 8.506624e-01 8.806217e-01 8.410164e-01 8.421512e-01 8.631454e-01
#>  [51,]    0 8.506624e-01 8.806217e-01 8.410164e-01 8.421512e-01 8.631454e-01
#>  [52,]    0 8.506624e-01 8.806217e-01 8.410164e-01 8.421512e-01 8.631454e-01
#>  [53,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [54,]    0 8.506624e-01 8.806217e-01 8.410164e-01 8.421512e-01 8.631454e-01
#>  [55,]    0 8.506624e-01 8.806217e-01 8.410164e-01 8.421512e-01 8.631454e-01
#>  [56,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [57,]    0 8.506624e-01 8.806217e-01 8.410164e-01 8.421512e-01 8.631454e-01
#>  [58,]    0 8.506624e-01 8.806217e-01 8.410164e-01 8.421512e-01 8.631454e-01
#>  [59,]    0 8.506624e-01 8.806217e-01 8.410164e-01 8.421512e-01 8.631454e-01
#>  [60,]    0 8.506624e-01 8.806217e-01 8.410164e-01 8.421512e-01 8.631454e-01
#>  [61,]    0 8.506624e-01 8.806217e-01 8.410164e-01 8.421512e-01 8.631454e-01
#>  [62,]    0 8.506624e-01 8.806217e-01 8.410164e-01 8.421512e-01 8.631454e-01
#>  [63,]    0 8.506624e-01 8.806217e-01 8.410164e-01 8.421512e-01 8.631454e-01
#>  [64,]    0 0.000000e+00 0.000000e+00 1.762069e-01 6.420037e-01 9.200000e-01
#>  [65,]    0 0.000000e+00 0.000000e+00 1.762069e-01 6.420037e-01 9.200000e-01
#>  [66,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [67,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [68,]    0 1.125773e-02 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [69,]    0 1.125773e-02 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [70,]    0 1.436283e-03 9.296757e-02 9.040309e-02 0.000000e+00 0.000000e+00
#>  [71,]    0 4.905232e-02 6.309385e-02 2.380985e-01 4.619029e-01 5.936285e-01
#>  [72,]    0 4.905232e-02 6.309385e-02 2.380985e-01 4.619029e-01 5.936285e-01
#>  [73,]    0 3.779459e-02 1.002005e-01 3.561138e-01 4.889615e-01 6.814289e-01
#>  [74,]    0 3.779459e-02 1.002005e-01 3.561138e-01 4.889615e-01 6.814289e-01
#>  [75,]    0 4.761604e-02 0.000000e+00 1.476954e-01 5.127963e-01 6.390050e-01
#>  [76,]    0 4.761604e-02 0.000000e+00 1.476954e-01 5.127963e-01 6.390050e-01
#>  [77,]    0 6.267632e-01 8.010133e-01 9.200000e-01 8.946304e-01 8.172916e-01
#>  [78,]    0 6.267632e-01 8.010133e-01 9.200000e-01 8.946304e-01 8.172916e-01
#>  [79,]    0 1.418867e-01 2.498266e-01 7.390936e-01 9.166276e-01 9.200000e-01
#>  [80,]    0 7.122362e-02 2.466536e-01 3.539526e-01 4.734115e-01 6.030128e-01
#>  [81,]    0 4.635844e-01 4.852976e-01 8.191787e-01 9.200000e-01 9.200000e-01
#>  [82,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [83,]    0 4.635844e-01 4.852976e-01 8.191787e-01 9.200000e-01 9.200000e-01
#>  [84,]    0 1.324514e-01 1.962905e-01 3.436576e-01 4.591051e-01 5.682879e-01
#>  [85,]    0 1.324514e-01 1.962905e-01 3.436576e-01 4.591051e-01 5.682879e-01
#>  [86,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [87,]    0 1.324514e-01 1.962905e-01 3.436576e-01 4.591051e-01 5.682879e-01
#>  [88,]    0 1.324514e-01 1.962905e-01 3.436576e-01 4.591051e-01 5.682879e-01
#>  [89,]    0 2.070918e-01 4.746776e-01 8.827136e-01 9.200000e-01 9.200000e-01
#>  [90,]    0 2.070918e-01 4.746776e-01 8.827136e-01 9.200000e-01 9.200000e-01
#>  [91,]    0 2.070918e-01 4.746776e-01 8.827136e-01 9.200000e-01 9.200000e-01
#>  [92,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [93,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [94,]    0 2.982798e-01 4.864368e-01 8.354023e-01 9.105192e-01 9.200000e-01
#>  [95,]    0 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00 0.000000e+00
#>  [96,]    0 2.982798e-01 4.864368e-01 8.354023e-01 9.105192e-01 9.200000e-01
#>  [97,]    0 2.982798e-01 4.864368e-01 8.354023e-01 9.105192e-01 9.200000e-01
#>  [98,]    0 2.982798e-01 4.864368e-01 8.354023e-01 9.105192e-01 9.200000e-01
#>  [99,]    0 2.982798e-01 4.864368e-01 8.354023e-01 9.105192e-01 9.200000e-01
#>