counts {rex}	R Documentation

Counts

Description

Functions to restrict a regex to a specific number

Usage

n_times(x, n, type = c("greedy", "lazy", "possessive"))

between(x, low, high, type = c("greedy", "lazy", "possessive"))

at_least(x, n, type = c("greedy", "lazy", "possessive"))

at_most(x, n, type = c("greedy", "lazy", "possessive"))

Arguments

x
A regex pattern.

n
An integer number

type
the type of match to perform.

There are three match types

greedy: match the longest string. This is the default matching type.

lazy: match the shortest string. This matches the shortest string from the same anchor point, not necessarily the shortest global string.

possessive: match and don't allow backtracking

low
An integer number for the lower limit.

high
An integer number for the upper limit.

Functions

n_times: x must occur exactly n times.

between: x must occur between low and high times.

at_least: x must occur at least n times.

at_most: x must occur at most n times.

See Also

Other rex: %or%(), capture(), character_class(), group(), lookarounds, not(), rex(), shortcuts, wildcards

[Package rex version 1.2.0 Index]