rex_mode {rex}	R Documentation

Toggles rex mode.

Description

While within rex mode, functions used within the rex function are attached, so one can get e.g. auto-completion within editors.

Usage

rex_mode()

[Package rex version 1.2.0 Index]